+++ core/domain/src/main/java/com/ilsangtech/ilsang/core/domain/QuestRepository.kt
@@ -1,11 +1,11 @@
 package com.ilsangtech.ilsang.core.domain
 
 import androidx.paging.PagingData
-import com.ilsangtech.ilsang.core.model.QuestType
 import com.ilsangtech.ilsang.core.model.quest.BannerQuest
 import com.ilsangtech.ilsang.core.model.quest.LargeRewardQuest
 import com.ilsangtech.ilsang.core.model.quest.PopularQuest
 import com.ilsangtech.ilsang.core.model.quest.QuestDetail
+import com.ilsangtech.ilsang.core.model.quest.QuestType
 import com.ilsangtech.ilsang.core.model.quest.RecommendedQuest
 import com.ilsangtech.ilsang.core.model.quest.TypedQuest
 import kotlinx.coroutines.flow.Flow
