+++ core/model/src/main/java/com/ilsangtech/ilsang/core/model/quest/QuestDetail.kt
@@ -1,6 +1,5 @@
 package com.ilsangtech.ilsang.core.model.quest
 
-import com.ilsangtech.ilsang.core.model.QuestType
 import com.ilsangtech.ilsang.core.model.RewardPoint
 import com.ilsangtech.ilsang.core.model.mission.Mission
