+++ core/model/src/main/java/com/ilsangtech/ilsang/core/model/quest/TypedQuest.kt
@@ -1,6 +1,5 @@
 package com.ilsangtech.ilsang.core.model.quest
 
-import com.ilsangtech.ilsang.core.model.QuestType
 import com.ilsangtech.ilsang.core.model.RewardPoint
 
 data class TypedQuest(
