+++ core/ui/src/main/java/com/ilsangtech/ilsang/core/ui/quest/QuestImageWithBadge.kt
@@ -8,7 +8,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.zIndex
-import com.ilsangtech.ilsang.core.model.QuestType
+import com.ilsangtech.ilsang.core.model.quest.QuestType
 
 @Composable
 fun QuestImageWithBadge(
