+++ feature/submit/src/main/java/com/ilsangtech/ilsang/feature/submit/model/SubmitQuestUiState.kt
@@ -1,7 +1,7 @@
 package com.ilsangtech.ilsang.feature.submit.model
 
-import com.ilsangtech.ilsang.core.model.QuestType
 import com.ilsangtech.ilsang.core.model.RewardPoint
+import com.ilsangtech.ilsang.core.model.quest.QuestType
 
 data class SubmitQuestUiState(
     val questImageId: String?,
